backend/requirements.txt
pandas
numpy
scikit-learn
joblib
matplotlib
seaborn
fastapi
uvicorn
python-dotenv
requests
pydantic
